+++ app/assets/javascripts/admin/addon/components/site-settings/string.hbs
@@ -1,7 +1,12 @@
 {{#if this.setting.textarea}}
   <Textarea @value={{this.value}} class="input-setting-textarea" />
 {{else if this.isSecret}}
-  <Input @type="password" @value={{this.value}} class="input-setting-string" />
+  <Input
+    @type="password"
+    @value={{this.value}}
+    class="input-setting-string"
+    autocomplete="new-password"
+  />
 {{else}}
   <TextField @value={{this.value}} @classNames="input-setting-string" />
 {{/if}}
