+++ app/assets/javascripts/admin/addon/templates/plugins-show-settings.hbs
@@ -1,1 +1,8 @@
-<div class="content-body admin-plugin-config-area__settings"></div>
+<div
+  class="content-body admin-plugin-config-area__settings admin-detail pull-left"
+>
+  <AdminPluginFilteredSiteSettings
+    @plugin={{@model.plugin}}
+    @settings={{@model.settings}}
+  />
+</div>
