+++ app/assets/javascripts/admin/addon/templates/site-settings.hbs
@@ -1,33 +1,8 @@
-<div class="admin-controls">
-  <div class="controls">
-    <div class="inline-form">
-      <DButton @action={{this.toggleMenu}} @icon="bars" class="menu-toggle" />
-      <TextField
-        @id="setting-filter"
-        @value={{this.filter}}
-        @placeholderKey="type_to_filter"
-        class="no-blur"
-      />
-      <DButton
-        @action={{this.clearFilter}}
-        @label="admin.site_settings.clear_filter"
-        id="clear-filter"
-        class="btn-default"
-      />
-    </div>
-  </div>
-
-  <div class="search controls">
-    <label>
-      <Input
-        @type="checkbox"
-        @checked={{this.onlyOverridden}}
-        class="toggle-overridden"
-      />
-      {{i18n "admin.settings.show_overriden"}}
-    </label>
-  </div>
-</div>
+<AdminSiteSettingsFilterControls
+  @onChangeFilter={{this.filterChanged}}
+  @showMenu={{true}}
+  @onToggleMenu={{this.toggleMenu}}
+/>
 
 <div class="admin-nav admin-site-settings-category-nav pull-left">
   <ul class="nav nav-stacked">
